+++ render/iswapchain.h
@@ -13,10 +13,10 @@ namespace Lightning
 			virtual ~ISwapChain() = default;
 			//preset the current back buffer to display
 			virtual bool Present() = 0;
-			virtual std::size_t GetSampleCount()const = 0;
+			virtual std::uint32_t GetSampleCount()const = 0;
 			virtual int GetSampleQuality()const = 0;
 			virtual RenderFormat GetRenderFormat()const = 0;
-			virtual std::size_t GetCurrentBackBufferIndex()const = 0;
+			virtual std::uint32_t GetCurrentBackBufferIndex()const = 0;
 			//get corresponding render target
 			virtual SharedRenderTargetPtr GetDefaultRenderTarget() = 0;
 		};
